refactor(router): правки после ревью
Jira issue: VEGA-596
